+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ### Changes
 * Added `kurtosis-engine-server` to the monorepo
+* Merged the `kurtosis-engine-sdk` & `kurtosis-core-sdk`
 
 # 0.49.4