+++ docs/changelog.md
@@ -1,5 +1,8 @@
 # TBD
 
+### Removals
+* Remove unused variables from Docker Kurtosis backend
+
 # 0.49.4
 
 ### Fixes