+++ docs/changelog.md
@@ -1,5 +1,7 @@
 # TBD
 
+# 0.49.5
+
 ### Changes
 * Added `kurtosis-engine-server` to the monorepo
 * Merged the `kurtosis-engine-sdk` & `kurtosis-core-sdk`